pyproject.toml
[project]
name = "deepseek-ocr-mps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "addict>=2.4.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "gradio>=3.36.1",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers==4.46.3",
]

[dependency-groups]
dev = ["mypy<2.0.0,>=1.8.0", "ruff<1.0.0,>=0.2.2", "huggingface_hub>=0.35.3"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
